moved ConnectionUtil class to util package, changed id variable type to Long

+++ src/main/java/mate/academy/dao/BookDao.java
@@ -1,6 +1,5 @@
 package mate.academy.dao;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 import mate.academy.model.Book;
@@ -8,11 +7,11 @@ import mate.academy.model.Book;
 public interface BookDao {
     Book create(Book book);
 
-    Optional<Book> findById(BigInteger id);
+    Optional<Book> findById(Long id);
 
     List<Book> findAll();
 
     Book update(Book book);
 
-    boolean deleteById(BigInteger id);
+    boolean deleteById(Long id);
 }
